Off-Platform Project: Stack Overflow Survey Trends
I didn't code this myself (it was an article), but I did meticulously research each cluster of code and mod it repeatedly to confirm I knew how it worked. Of particular hassle was the DataFrame.loc = value code, which I didn't know could create extra columns in the DF until this exercise.